--- packages/apex-lsp-testbed/test/nodeserver-benchmark-results.json
@@ -1,225 +1,0 @@
-{
-  "NodeServer LSP Complete Workflow": {
-    "name": "NodeServer LSP Complete Workflow",
-    "options": {
-      "async": false,
-      "defer": true,
-      "delay": 0.005,
-      "initCount": 1,
-      "maxTime": 6,
-      "minSamples": 2,
-      "minTime": 2
-    },
-    "async": false,
-    "defer": true,
-    "delay": 0.005,
-    "initCount": 1,
-    "maxTime": 6,
-    "minSamples": 2,
-    "minTime": 2,
-    "id": 1,
-    "stats": {
-      "moe": 0.0006895429180339243,
-      "rme": 0.09639163716940667,
-      "sem": 0.0002167011056046274,
-      "deviation": 0.0004334022112092548,
-      "mean": 0.7153555414999999,
-      "sample": [
-        0.714800486,
-        0.7158162776666667,
-        0.7155418053333333,
-        0.715263597
-      ],
-      "variance": 1.878374766810715e-7
-    },
-    "times": {
-      "cycle": 2.1460666244999995,
-      "elapsed": 9.404,
-      "period": 0.7153555414999999,
-      "timeStamp": 1752251130209
-    },
-    "_timerId": {
-      "_idleTimeout": 5,
-      "_idlePrev": null,
-      "_idleNext": null,
-      "_idleStart": 8420,
-      "_repeat": null,
-      "_destroyed": true
-    },
-    "events": {
-      "complete": []
-    },
-    "running": false,
-    "count": 3,
-    "cycles": 2,
-    "hz": 1.3979062745542457
-  },
-  "NodeServer LSP textDocument/didOpen Id: 8": {
-    "name": "NodeServer LSP textDocument/didOpen Id: 8",
-    "options": {
-      "async": false,
-      "defer": true,
-      "delay": 0.005,
-      "initCount": 1,
-      "maxTime": 8,
-      "minSamples": 2,
-      "minTime": 2
-    },
-    "async": false,
-    "defer": true,
-    "delay": 0.005,
-    "initCount": 1,
-    "maxTime": 8,
-    "minSamples": 2,
-    "minTime": 2,
-    "id": 2,
-    "stats": {
-      "moe": 0.00035498682939521836,
-      "rme": 0.17272912231773635,
-      "sem": 0.00012787709992623142,
-      "deviation": 0.0002859418882005868,
-      "mean": 0.20551649,
-      "sample": [
-        0.20590411670000003,
-        0.2053797625,
-        0.2053921125,
-        0.2057126791,
-        0.20519377919999998
-      ],
-      "variance": 8.176276342771689e-8
-    },
-    "times": {
-      "cycle": 2.0551649,
-      "elapsed": 10.563,
-      "period": 0.20551649,
-      "timeStamp": 1752251139623
-    },
-    "_timerId": {
-      "_idleTimeout": 5,
-      "_idlePrev": null,
-      "_idleNext": null,
-      "_idleStart": 17834,
-      "_repeat": null,
-      "_destroyed": true
-    },
-    "events": {
-      "complete": []
-    },
-    "running": false,
-    "count": 10,
-    "cycles": 2,
-    "hz": 4.865789601603258
-  },
-  "NodeServer LSP textDocument/documentSymbol Id: 9": {
-    "name": "NodeServer LSP textDocument/documentSymbol Id: 9",
-    "options": {
-      "async": false,
-      "defer": true,
-      "delay": 0.005,
-      "initCount": 1,
-      "maxTime": 8,
-      "minSamples": 2,
-      "minTime": 2
-    },
-    "async": false,
-    "defer": true,
-    "delay": 0.005,
-    "initCount": 1,
-    "maxTime": 8,
-    "minSamples": 2,
-    "minTime": 2,
-    "id": 3,
-    "stats": {
-      "moe": 0.0008825227541021818,
-      "rme": 0.4196219038609817,
-      "sem": 0.0003179116549359445,
-      "deviation": 0.0007108720712762285,
-      "mean": 0.21031379582,
-      "sample": [
-        0.20917261250000002,
-        0.2102866167,
-        0.2103226625,
-        0.21104524160000002,
-        0.21074184580000002
-      ],
-      "variance": 5.053391017205553e-7
-    },
-    "times": {
-      "cycle": 2.1031379582,
-      "elapsed": 10.811,
-      "period": 0.21031379582,
-      "timeStamp": 1752251150194
-    },
-    "_timerId": {
-      "_idleTimeout": 5,
-      "_idlePrev": null,
-      "_idleNext": null,
-      "_idleStart": 28405,
-      "_repeat": null,
-      "_destroyed": true
-    },
-    "events": {
-      "complete": []
-    },
-    "running": false,
-    "count": 10,
-    "cycles": 2,
-    "hz": 4.754799827092008
-  },
-  "NodeServer LSP textDocument/didClose Id: 12": {
-    "name": "NodeServer LSP textDocument/didClose Id: 12",
-    "options": {
-      "async": false,
-      "defer": true,
-      "delay": 0.005,
-      "initCount": 1,
-      "maxTime": 8,
-      "minSamples": 2,
-      "minTime": 2
-    },
-    "async": false,
-    "defer": true,
-    "delay": 0.005,
-    "initCount": 1,
-    "maxTime": 8,
-    "minSamples": 2,
-    "minTime": 2,
-    "id": 4,
-    "stats": {
-      "moe": 0.0008057184709311657,
-      "rme": 0.19758979996327397,
-      "sem": 0.00029024440595503087,
-      "deviation": 0.0006490062218044939,
-      "mean": 0.40777331172000003,
-      "sample": [
-        0.40733184180000004,
-        0.4086182,
-        0.4083219168,
-        0.4074020916,
-        0.4071925084
-      ],
-      "variance": 4.212090759409439e-7
-    },
-    "times": {
-      "cycle": 2.0388665586,
-      "elapsed": 10.686,
-      "period": 0.40777331172000003,
-      "timeStamp": 1752251161013
-    },
-    "_timerId": {
-      "_idleTimeout": 5,
-      "_idlePrev": null,
-      "_idleNext": null,
-      "_idleStart": 39224,
-      "_repeat": null,
-      "_destroyed": true
-    },
-    "events": {
-      "complete": []
-    },
-    "running": false,
-    "count": 5,
-    "cycles": 2,
-    "hz": 2.4523429348084846
-  }
-}
