+++ packages/apex-parser-ast/README.md
@@ -6,10 +6,10 @@ This package is the foundational component of the Apex Language Server responsib
 
 The package is built around a few core concepts that work together to turn raw source code into a meaningful data model.
 
--   **`CompilerService`**: The main entry point and orchestrator for the parsing process. It manages the entire pipeline, from receiving the source code to returning a final compilation result.
--   **`SymbolTable`**: The primary data structure produced by this package. It's a hierarchical tree of `SymbolScope` objects that mirrors the lexical scoping of the source code (file > class > method > block). It contains every `ApexSymbol` found in the file and allows for efficient lookup.
--   **`ApexSymbol`**: A generic data structure representing a single named entity in the code, such as a class, interface, method, property, or variable. Specialized versions (`TypeSymbol`, `MethodSymbol`, `VariableSymbol`) extend this base structure to hold relevant metadata.
--   **Parser Listeners**: Classes that subscribe to events emitted by the ANTLR parser as it traverses the parse tree. The key listener is the `ApexSymbolCollectorListener`, which is responsible for building the `SymbolTable`.
+- **`CompilerService`**: The main entry point and orchestrator for the parsing process. It manages the entire pipeline, from receiving the source code to returning a final compilation result.
+- **`SymbolTable`**: The primary data structure produced by this package. It's a hierarchical tree of `SymbolScope` objects that mirrors the lexical scoping of the source code (file > class > method > block). It contains every `ApexSymbol` found in the file and allows for efficient lookup.
+- **`ApexSymbol`**: A generic data structure representing a single named entity in the code, such as a class, interface, method, property, or variable. Specialized versions (`TypeSymbol`, `MethodSymbol`, `VariableSymbol`) extend this base structure to hold relevant metadata.
+- **Parser Listeners**: Classes that subscribe to events emitted by the ANTLR parser as it traverses the parse tree. The key listener is the `ApexSymbolCollectorListener`, which is responsible for building the `SymbolTable`.
 
 ## Architecture and Workflow
 
@@ -17,8 +17,8 @@ The package uses a classic compiler front-end architecture based on the ANTLR pa
 
 1.  **Orchestration**: A consumer (like a language server) invokes the `CompilerService.compile()` method, passing in the Apex source code and a listener instance.
 2.  **ANTLR Pipeline**: The `CompilerService` sets up the ANTLR pipeline:
-    -   The source code is fed into an `ApexLexer`, which performs lexical analysis, breaking the string into a sequence of tokens.
-    -   The tokens are passed to an `ApexParser`, which builds a concrete Parse Tree based on the Apex grammar.
+    - The source code is fed into an `ApexLexer`, which performs lexical analysis, breaking the string into a sequence of tokens.
+    - The tokens are passed to an `ApexParser`, which builds a concrete Parse Tree based on the Apex grammar.
 3.  **Tree Traversal**: A `ParseTreeWalker` traverses the generated parse tree node by node.
 4.  **Event-Driven Analysis**: As the walker visits each node (e.g., a class declaration), it triggers the corresponding method on the provided listener (e.g., `enterClassDeclaration`).
 5.  **Symbol Table Construction**: The `ApexSymbolCollectorListener` responds to these events. When it enters a new scope (like a class), it tells the `SymbolTable` to `enterScope()`. When it finds a symbol (like a method), it creates an `ApexSymbol` and adds it to the current scope. When it leaves a scope, it calls `exitScope()`.
@@ -26,15 +26,15 @@ The package uses a classic compiler front-end architecture based on the ANTLR pa
 
 ## Directory Structure
 
--   `src/parser/`: Contains the ANTLR-driven parsing infrastructure.
-    -   `compilerService.ts`: The high-level `CompilerService` that orchestrates the process.
-    -   `listeners/`: Home to the various parser listeners, including `ApexSymbolCollectorListener`, which builds the symbol table, and `ApexErrorListener`, which collects syntax errors.
--   `src/types/`: Defines the core data model for the AST.
-    -   `symbol.ts`: Defines the `ApexSymbol` hierarchy and the `SymbolTable` / `SymbolScope` classes.
-    -   `typeInfo.ts`: Defines structures for representing type information.
--   `src/semantics/`: Contains modules for performing semantic analysis and validation on the AST after it's built.
--   `src/generated/`: Contains the lexer and parser code automatically generated by ANTLR from a grammar file (not present in this repo, but a standard part of the build process).
--   `src/index.ts`: The public API entry point for the package, exporting all consumable classes and types.
+- `src/parser/`: Contains the ANTLR-driven parsing infrastructure.
+  - `compilerService.ts`: The high-level `CompilerService` that orchestrates the process.
+  - `listeners/`: Home to the various parser listeners, including `ApexSymbolCollectorListener`, which builds the symbol table, and `ApexErrorListener`, which collects syntax errors.
+- `src/types/`: Defines the core data model for the AST.
+  - `symbol.ts`: Defines the `ApexSymbol` hierarchy and the `SymbolTable` / `SymbolScope` classes.
+  - `typeInfo.ts`: Defines structures for representing type information.
+- `src/semantics/`: Contains modules for performing semantic analysis and validation on the AST after it's built.
+- `src/generated/`: Contains the lexer and parser code automatically generated by ANTLR from a grammar file (not present in this repo, but a standard part of the build process).
+- `src/index.ts`: The public API entry point for the package, exporting all consumable classes and types.
 
 ## Overview
 
@@ -144,7 +144,7 @@ The parser captures inheritance relationships between types:
   All references to Babel have been removed from the project. The project now uses `ts-jest` exclusively for testing.
 
 - **TypeScript Improvements:**  
-  Explicit types have been added to test files to resolve TypeScript errors. For example, in `apex-lsp-testbed/test/performance/lsp-benchmarks.web.test.ts`, variables and parameters now have explicit `any` types.
+  Explicit types have been added to test files to resolve TypeScript errors. For example, in `apex-lsp-testbed/test/performance/lsp-benchmarks.test.ts`, variables and parameters now have explicit `any` types.
 
 - **Jest Configuration:**  
   Jest configurations have been streamlined. Each package now uses a single Jest configuration file (`jest.config.cjs`), and the `"jest"` key has been removed from `package.json` files to avoid conflicts.
